+++ jsonapi_demo_cli/jsonapi_demo_cli/__init__.py
@@ -1,0 +1,2 @@
+from .oauth2_session import OAuth2Session
+from .cli import main