+++ jsonapi_demo_cli/jsonapi_demo_cli/__main__.py
@@ -1,0 +1,4 @@
+from jsonapi_demo_cli import main
+
+if __name__ == '__main__':
+    main()