+++ myapp/__init__.py
@@ -1,8 +1,8 @@
 __title__ = 'myapp'
-__version__ = '0.2.3'
+__version__ = '0.2.4'
 __author__ = 'Alan Crosswell'
 __license__ = 'CCbyncsa4.0'
-__copyright__ = '(c) 2018 The Trustees of Columbia University in the City of New York'
+__copyright__ = '(c) 2018-2019 The Trustees of Columbia University in the City of New York'
 
 # Version synonym
 VERSION = __version__